+++ src/agent.cpp
@@ -232,7 +232,7 @@ bool Agent::randomAgentNeedsHospital(double agentNeedsHospital[18]){
     return agentNeedsHospital[getAgentAgeGroup()] >= ((double) rand() / (RAND_MAX)) && (getSeverity() != INCUBATION);
 }
 
-void Agent::agentIncubationCheck(short agentIncubationTime[18]){
+void Agent::agentIncubationCheck(int agentIncubationTime[18]){
     //this is multiplied by the currTime step because this time will increment every time step instead of every day
     if(timeIncubating < ((int)agentIncubationTime[getAgentAgeGroup()])){
         timeIncubating = 0;
@@ -242,7 +242,7 @@ void Agent::agentIncubationCheck(short agentIncubationTime[18]){
     }
 }
 
-void Agent::agentInfectedCheck(short agentRecoveryTime[18]){
+void Agent::agentInfectedCheck(int agentRecoveryTime[18]){
     if(timeInfected < ((int)agentRecoveryTime[getAgentAgeGroup()])){
         timeInfected = 0;
         recoverAgent();
