+++ src/hospital.cpp
@@ -67,7 +67,7 @@ void Hospital::increaseIcuCount(Agent* agentToAdd) {
     indicateOverflow(); 
 }
 
-void Hospital::HospitalTimeStep(double timestep, short agentRecoveryTime[18], double agentDeathChance[18], double agentChanceOfICU[18]) {
+void Hospital::HospitalTimeStep(double timestep, int agentRecoveryTime[18], double agentDeathChance[18], double agentChanceOfICU[18]) {
     //loop through all the beds in the icu
     string sirResponse;
     Agent *currAgent;
